+++ CHANGELOG.md
@@ -4,10 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [0.0.0] - 2024-09-12
 
 ### Added
 
+- Create tests for `Helpers` class.
+- Add `Helpers` class with `getAppInfoArray` and `getAppInfo` methods.
+- Initial project structure.
 - New changelog file.
 
+[Unreleased]: git diff staging..dev
 [0.0.0]: git log v0.0.0